believe CRUD Controller is working

+++ Models/Seed.cs
@@ -6,29 +6,29 @@ public static class Seed
 {
     public static void InitializeDev(DB db)
     {   
-        // Look for any Posts.
-        // if (db.Cards.Any() || db.CardLists.Any())
+        // Look for any Tweets.
+        if (db.Tweets.Any() || db.TweetList.Any())
         db.Database.EnsureDeleted(); // delete then, ...
         db.Database.EnsureCreated(); // create the tables!! 
         
-        CardList todo = new CardList { Summary="Todo items", Cards = new List<Card>() };
+        TweetList todo = new TweetList { Summary="Tweets", Tweets = new List<Tweet>() };
 
         for(var i = 0; i < 10; i++){
-            todo.Cards.Add(
-                new Card { Title = $"Test Card {i}", Content = $"Test Content {i}",  }
+            todo.Tweets.Add(
+                new Tweet { Content = $"Tweet Content {i}",  }
             );
         }
 
-        db.CardLists.Add(todo);
+         db.TweetList.Add(todo);
 
         db.SaveChanges();
 
         // print out the tables 
-        Console.WriteLine("----------CARDS SEEDED-------------");
+        Console.WriteLine("----------TWEETS SEEDED-------------");
     }
 
-    public static void InitializeProd(DB db){
-        db.Database.EnsureCreated();
-        // db.Database.Migrate();
-    }
+    // public static void InitializeProd(DB db){
+    //     db.Database.EnsureCreated();
+    //      db.Database.Migrate();
+    // }
 }
